Move most youtube specific tests into the test_youtube.py file.
