+++ tests/test_youtube.py
@@ -4,6 +4,9 @@ from unittest.mock import patch
 
 import pytest
 
+from ricecooker.classes.files import YouTubeSubtitleFile
+from ricecooker.utils.youtube import get_language_with_alpha2_fallback
+from ricecooker.utils.youtube import is_youtube_subtitle_file_supported_language
 from ricecooker.utils.youtube import YouTubePlaylistUtils
 from ricecooker.utils.youtube import YouTubeVideoUtils
 
@@ -78,3 +81,62 @@ def test_youtube_playlist_cache(youtube_playlist_cache):
         )
     )
     assert playlist_info and os.path.exists(playlist_cache_filepath)
+
+
+@pytest.fixture
+def subtitles_langs_internal():
+    return ["en", "es", "pt-BR"]
+
+
+@pytest.fixture
+def subtitles_langs_pycountry_mappable():
+    return ["zu"]
+
+
+@pytest.fixture
+def subtitles_langs_youtube_custom():
+    return ["iw", "zh-Hans", "pt-BR"]
+
+
+@pytest.fixture
+def subtitles_langs_unsupported():
+    return ["sgn", "zzzza", "ab-dab", "bbb-qqq"]
+
+
+@pytest.mark.skipif(True, reason="Requires connecting to youtube.")
+def test_youtubesubtitle_process_file(youtube_video_with_subs_dict):
+    youtube_id = youtube_video_with_subs_dict["youtube_id"]
+    lang = youtube_video_with_subs_dict["subtitles_langs"][0]
+    sub_file = YouTubeSubtitleFile(youtube_id=youtube_id, language=lang)
+    filename = sub_file.process_file()
+    assert filename is not None, "Processing YouTubeSubtitleFile file failed"
+    assert filename.endswith(".vtt"), "Wrong extension for video subtitles"
+    assert not filename.endswith("." + lang + ".vtt"), "Lang code in extension"
+
+
+def test_is_youtube_subtitle_file_supported_language(
+    subtitles_langs_internal,
+    subtitles_langs_pycountry_mappable,
+    subtitles_langs_youtube_custom,
+):
+    for lang in subtitles_langs_internal:
+        assert is_youtube_subtitle_file_supported_language(lang), "should be supported"
+        lang_obj = get_language_with_alpha2_fallback(lang)
+        assert lang_obj is not None, "lookup should return Language object"
+    for lang in subtitles_langs_pycountry_mappable:
+        assert is_youtube_subtitle_file_supported_language(lang), "should be supported"
+        lang_obj = get_language_with_alpha2_fallback(lang)
+        assert lang_obj is not None, "lookup should return Language object"
+    for lang in subtitles_langs_youtube_custom:
+        assert is_youtube_subtitle_file_supported_language(lang), "should be supported"
+        lang_obj = get_language_with_alpha2_fallback(lang)
+        assert lang_obj is not None, "lookup should return Language object"
+
+
+def test_is_youtube_subtitle_file_unsupported_language(subtitles_langs_unsupported):
+    for lang in subtitles_langs_unsupported:
+        assert not is_youtube_subtitle_file_supported_language(
+            lang
+        ), "should not be supported"
+        lang_obj = get_language_with_alpha2_fallback(lang)
+        assert lang_obj is None, "lookup should fail"
